Refactored application insights and log analytics into its own package
